+++ CHANGELOG.md
@@ -1072,3 +1072,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix WhatMod not working for powers on the player (kiooeht)
 * Improve visual of hand sizes greater than 10 (kiooeht)
 * Add run info panel to main menu, inspired by sts2 preview (kiooeht)
+* Fix Watcher visual bug on some aspect ratios (kiooeht)